# Saas-Project/backend/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS UMKM Backend API
  description: |
    API documentation for Indonesian SME licensing and business management platform.
    This API provides endpoints for user management, company registration, license 
    management, and business operations.
  version: 0.1.0
  contact:
    name: SaaS UMKM Team
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api-staging.saas-umkm.com/api
    description: Staging server
  - url: https://api.saas-umkm.com/api
    description: Production server
...
paths:
...
  /companies:
    post:
      summary: Create a new company
      description: Creates a new company for the authenticated user
      operationId: createCompany
      tags:
        - Companies
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
      responses:
        '201':
          description: Company created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List companies
      description: List companies owned by the authenticated user or all companies for admin
      operationId: listCompanies
      tags:
        - Companies
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: Companies retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompanyResponse'
                  pagination:
                    $ref: '#/components/schemas/PaginationMetadata'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /companies/{id}:
    get:
      summary: Get company by ID
      description: Retrieves a company by its ID
      operationId: getCompany
      tags:
        - Companies
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Company ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Company retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user doesn't have access to this company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Company not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update company
      description: Updates a company's information
      operationId: updateCompany
      tags:
        - Companies
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Company ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
      responses:
        '200':
          description: Company updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user doesn't have access to this company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Company not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete company
      description: Deletes a company
      operationId: deleteCompany
      tags:
        - Companies
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Company ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Company deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user doesn't have access to this company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Company not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    CompanyRequest:
      type: object
      required:
        - name
        - business_type
        - address
      properties:
        name:
          type: string
          example: Toko Jaya Abadi
        business_type:
          type: string
          example: retail
        address:
          type: string
          example: Jl. Pahlawan No. 123, Jakarta
        phone:
          type: string
          example: '+62212345678'
        tax_id:
          type: string
          example: '09.123.456.7-123.000'
        registration_number:
          type: string
          example: 'AHU-12345.AH.01.01.TAHUN2024'
    CompanyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        owner_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        name:
          type: string
          example: Toko Jaya Abadi
        business_type:
          type: string
          example: retail
        address:
          type: string
          example: Jl. Pahlawan No. 123, Jakarta
        phone:
          type: string
          example: '+62212345678'
        tax_id:
          type: string
          example: '09.123.456.7-123.000'
        registration_number:
          type: string
          example: 'AHU-12345.AH.01.01.TAHUN2024'
        verified:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: '2025-07-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-07-01T12:00:00Z'
    PaginationMetadata:
      type: object
      properties:
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
        has_more:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid input
        message:
          type: string
          example: Email is already in use
        status:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: '2025-07-01T12:00:00Z'
  parameters:
    limitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    offsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT